requirements.txt

python-telegram-bot
persiantools
urllib3